controllers added to be

+++ backend/controllers/users/delete.users.js
@@ -2,13 +2,19 @@ import {db} from "../../config/connect.js"
 //delete a user
 export const deleteUser = (req,res) =>{
     const userId = req.params.id
-    const q = "DELETE FROM users WHERE id = ?"
-    db.query(q,userId,(error,data)=>{
-        if (error) {
-            return res.json(error)
-        } else {
-            return res.json(data)            
-        }
-    })
+    if (req.user.id == userId) {
+        const q = "DELETE FROM users WHERE id = ?"
+        db.query(q,userId,(error,data)=>{
+            if (error) {
+                return res.json(error)
+            } else {
+                return res.json(data)            
+            }
+        })
+    }else{
+        return res.json("You are not allowed to delete it.")
+    }
+
+ 
 }
